Handle globs and multiple arguments in cli
This implementation interprets globs very strictly. For example, if you
have two test files in an `examples/` directory, then running the cli
with `examples/` as an argument will not run everything in that
directory but instead try to require `examples/` as a test. Instead, you
need to use a 'proper' glob like `examples/*`.

Another pass at this should look at error handling and more intelligent
interpretation of arguments.