Remove messy markStartLine maneuver in test definitions
It's necessary for tests to communicate their start line up the stack,
rather than encapsulating that data within the test--this is because the
wrapping describe needs to provide each test's start line as the end
line to the previous test.

Previously this was accomplished messily, by stashing the start line as
a property on the test function. It made the `define-test` function
harder to read, let alone the usual risks of writing properties to
objects you don't control.

Instead, this commit has `it` and `describe` return small data
structures rather than functions. This data structure includes all the
information that the next level up needs: the test function (aka the
"definition"), and the test's start line. No clutter necessary in the
`define-test` function, and hopefully more straightforward for readers.