Extract application of line number scope in describe
Trying to make the core export of `describe.js` more readable. There's a
little trickery involved in creating a staggered zip of the test list to
make it easy to call the definition function with its own start line as
the true start line, and the following test's start line as the end
line. Reasoning through that shouldn't be necessary to reason about the
meat of `describe`, so hide it further down in the file.

(I considered pulling `scopeAllTestsByLineNumber` into a separate module
entirely, but talked myself out of it...for now.)